Overall works -- but not the saver
The saver doesn't know how to load new files -- perhaps that should be something attached to the individual object
