+++ GAlgorithm/__init__.py
@@ -3,3 +3,5 @@
 from .bases import AbstractIndividual
 from .bases import Population
 from .geneticalgorithm import GeneticAlgorithm
+
+from .saver import load, dump
