+++ GAlgorithm/geneticalgorithm.py
@@ -3,109 +3,7 @@ import random
 
 from sklearn.naive_bayes import GaussianNB
 
-from bases import City, Route, Population
-
-
-def crossover(parents):
-    '''Crosses the parents over to create children'''
-    children:Route = []
-    for i, individual in enumerate(parents.individuals):
-        children.append(cross(individual, parents.individuals[-i-1]))
-    return Population(children)
-
-
-def cross(p1, p2):
-    '''Crosses parent1 with parent2
-    Note that all cities must be represented'''
-    
-    child = []
-
-    r1 = random.randint(0, len(p1.genes))
-    r2 = random.randint(0, len(p1.genes))
-
-    start = min(r1, r2)
-    stop = max(r1, r2)
-
-    cross_segment = p1.genes[start:stop]
-    needed_cities = [city for city in p1.genes if city not in cross_segment]
-
-    child.extend(cross_segment)
-
-    for _ in range(len(needed_cities)):
-        r = random.randint(0, len(needed_cities)-1)
-        child.append(needed_cities.pop(r))
-
-    return Route(child)
-
-
-def mutate(children, chance):
-    '''Mutates the children, given the chance
-    children: list of route children
-    chance: chance of mutation btw 0 and 1'''
-
-    for child in children.individuals:
-        mutate_2(child, chance)
-    return Population(children.individuals)
-
-
-def mutate_1(child, chance):
-    '''Mutates each gene in child with a chance of chance'''
-    for i in range(len(child.genes)):
-        if random.random() < chance:
-            r = random.randint(0, len(child.genes)-1)
-            child.genes[i], child.genes[r] = child.genes[r], child.genes[i]
-            child._fitness = None
-
-
-def mutate_2(child, chance):
-    '''Mutates each child with a chance of self.chance'''
-    if random.random() < chance:
-        A = random.randint(0, len(child.genes)-1)
-        B = random.randint(0, len(child.genes)-1)
-        child.genes[A], child.genes[B] = child.genes[B], child.genes[A]
-        child._fitness = None
-
-
-def select(population, tourny_size:int=2):
-    '''Selects a parent population from the population
-    Uses a tournament to select parents'''
-    parents = []
-    for _ in population.individuals:
-        winner = population.random_individual()
-        for _ in range(1, tourny_size):
-            competitor = population.random_individual()
-            if competitor.fitness > winner.fitness:
-                winner = competitor
-        parents.append(winner)
-    return Population(parents)
-
-
-def stochastic_acceptance_selection(population):
-    '''Selects parents based on stochastic acceptance.
-    1) Randomly select individual
-    2) Accept selection with probability fi/fm
-    where fm = maximum population fitness'''
-    max_fitness = population.max_fitness
-    min_fitness = population.min_fitness
-    parents = []
-    complete = False
-    while not complete:
-        individual = population.random_individual()
-        probality = (individual.fitness - min_fitness) / max_fitness
-        if random.random() <= probality:
-            parents.append(individual)
-            if len(parents) == len(population.individuals):
-                complete = True
-    return Population(parents)
-
-
-def evolve(population, tourny_size, mutation_rate):
-    '''Evolves the population to the next generation
-    Returns: a Population of the new generation'''
-    parents = select(population, tourny_size)
-    children1 = crossover(parents)
-    children2 = mutate(children1, mutation_rate)
-    return children2
+from .bases import Population
 
 
 def cull(keep:int, *args:Population):
@@ -124,6 +22,19 @@ class GeneticAlgorithm:
         self.tourny_size = tourny_size
         self.f_eval_max = f_eval_max
         self.reset()
+
+        self.select = lambda *args: None
+        self.crossover = lambda *args: None
+        self.mutate = lambda *args: None
+        
+
+    def evolve(self, population, tourny_size, mutation_rate):
+        '''Evolves the population to the next generation
+        Returns: a Population of the new generation'''
+        parents = self.select(population, tourny_size)
+        children1 = self.crossover(parents)
+        children2 = self.mutate(children1, mutation_rate)
+        return children2
 
 
     def reset(self):
@@ -153,7 +64,7 @@ class GeneticAlgorithm:
             self.pop_history.append(population)
 
             # evolve population
-            new_population = evolve(population, self.tourny_size, self.mutation_rate)
+            new_population = self.evolve(population, self.tourny_size, self.mutation_rate)
 
 
     def run_with_ml(self):
@@ -195,7 +106,7 @@ class GeneticAlgorithm:
             dev = population.get_standard_deviation()
             # print(f'Population standard deviation: {dev}')
             if dev < 5e-6 or no_improvement_count > 2:
-                new_population = mutate(population.copy(), 1)
+                new_population = self.mutate(population.copy(), 1)
                 # break the loop if the we have had 5 instances of low deviation
                 low_dev_count += 1
                 if low_dev_count > 3:
@@ -204,7 +115,7 @@ class GeneticAlgorithm:
                 print('just mutated this time')
             else:
                 low_dev_count = 0
-                new_population = evolve(population, self.tourny_size, self.mutation_rate)
+                new_population = self.evolve(population, self.tourny_size, self.mutation_rate)
 
             # classify
             good_pop, bad_pop = self.classify(new_population)
